+++ peaksMCP/workflows/__init__.py
@@ -2,11 +2,10 @@
 
 from peaksMCP.pxt_utils.metadata import read_meta
 
-from .publication import publication_grid, validate_arpes_metadata
+from .publication import validate_arpes_metadata
 from .slice_view import show_mapping_slice
 
 __all__ = [
-    "publication_grid",
     "validate_arpes_metadata",
     "read_meta",
     "show_mapping_slice",